No more kotlin!
A whole new project in java instead of kotlin.
The features are quite the same - time that updates every minute in the main activity, you can choose a date for the reminder and the date appeares in the list.

The kotlin project is backed up in a folder in the project.